+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/dotgenerator/Shared.scala
@@ -1,67 +1,12 @@
 package io.shiftleft.semanticcpg.dotgenerator
 
 import io.shiftleft.codepropertygraph.generated.nodes
-import overflowdb.Node
-import overflowdb.traversal._
 import io.shiftleft.semanticcpg.language._
 import io.shiftleft.semanticcpg.utils.MemberAccess
 
 object Shared {
 
   case class Edge(src: nodes.StoredNode, dst: nodes.StoredNode, label: String = "")
-
-  def dotGraph(method: nodes.Method,
-               expand: nodes.StoredNode => Iterator[Edge],
-               cfgNodeShouldBeDisplayed: Node => Boolean,
-  ): String = {
-    val sb = Shared.namedGraphBegin(method)
-    sb.append(nodesAndEdges(method, expand, cfgNodeShouldBeDisplayed).mkString("\n"))
-    Shared.graphEnd(sb)
-  }
-
-  private def nodesAndEdges(methodNode: nodes.Method,
-                            expand: nodes.StoredNode => Iterator[Edge],
-                            cfgNodeShouldBeDisplayed: Node => Boolean): List[String] = {
-    val vertices = methodNode.start.cfgNode.l ++ List(methodNode, methodNode.methodReturn) ++ methodNode.parameter.l
-    val verticesToDisplay = vertices.filter(cfgNodeShouldBeDisplayed)
-
-    def edgesToDisplay(srcNode: nodes.StoredNode, visited: List[nodes.StoredNode] = List()): List[Edge] = {
-      if (visited.contains(srcNode)) {
-        List()
-      } else {
-        val children = expand(srcNode).filter(x => vertices.contains(x.dst))
-        val (visible, invisible) = children.partition(x => cfgNodeShouldBeDisplayed(x.dst))
-        visible.toList ++ invisible.toList.flatMap { n =>
-          edgesToDisplay(n.dst, visited ++ List(srcNode)).map(y => Edge(srcNode, y.dst))
-        }
-      }
-    }
-
-    val edges = verticesToDisplay.map { v =>
-      edgesToDisplay(v)
-    }
-
-    val allIdsReferencedByEdges = edges.flatten.flatMap { edge =>
-      Set(edge.src.id, edge.dst.id)
-    }
-
-    val nodeStrings = verticesToDisplay.map { node =>
-      if (allIdsReferencedByEdges.contains(node.id)) {
-        s""""${node.id}" [label = "${Shared.stringRepr(node)}" ]""".stripMargin
-      } else {
-        ""
-      }
-    }
-
-    val edgeStrings = edges.flatMap { edges: List[Edge] =>
-      edges.map(
-        edge =>
-          s"""  "${edge.src.id}" -> "${edge.dst.id}" """ +
-            Some(s""" [ label = "${escape(edge.label)}"] """).filter(_ => edge.label != "").getOrElse(""))
-    }
-
-    nodeStrings ++ edgeStrings
-  }
 
   def namedGraphBegin(root: nodes.AstNode): StringBuilder = {
     val sb = new StringBuilder
@@ -89,23 +34,28 @@ object Shared {
 
   def toCfgNode(node: nodes.StoredNode): nodes.CfgNode = {
     node match {
-      case node: nodes.Identifier => node.parentExpression.get
-      case node: nodes.MethodRef  => node.parentExpression.get
-      case node: nodes.Literal    => node.parentExpression.get
-
+      case node: nodes.Identifier        => node.parentExpression.get
+      case node: nodes.MethodRef         => node.parentExpression.get
+      case node: nodes.Literal           => node.parentExpression.get
       case node: nodes.MethodParameterIn => node.method
-
       case node: nodes.MethodParameterOut =>
         node.method.methodReturn
-
       case node: nodes.Call if MemberAccess.isGenericMemberAccessName(node.name) =>
         node.parentExpression.get
-
       case node: nodes.Call         => node
       case node: nodes.ImplicitCall => node
       case node: nodes.MethodReturn => node
       case node: nodes.Expression   => node
     }
+  }
+
+  def nodeToDot(node: nodes.StoredNode): String = {
+    s""""${node.id}" [label = "${Shared.stringRepr(node)}" ]""".stripMargin
+  }
+
+  def edgeToDot(edge: Edge): String = {
+    s"""  "${edge.src.id}" -> "${edge.dst.id}" """ +
+      Some(s""" [ label = "${Shared.escape(edge.label)}"] """).filter(_ => edge.label != "").getOrElse("")
   }
 
   def escape(str: String): String = {
@@ -116,11 +66,5 @@ object Shared {
     sb.append("\n}\n")
     sb.toString
   }
-
-}
-
-abstract class CfgBasedDotGenerator {
-
-  protected def expand(node: nodes.CfgNode): java.util.Iterator[nodes.StoredNode]
 
 }
